Projects/Google-Agents/paint-planner/paint-agent-service/paint_agent/tools.py
from dotenv import load_dotenv
import os
from google.adk.tools import ToolContext

load_dotenv()


def set_session_value(tool_context: ToolContext, key: str, value: str):
    tool_context.state[key] = value
    return {"key": key, "value": value}
