Projects/Google-Agents/paint-planner/paint-agent-service/requirements.txt
bs4==0.0.2
beautifulsoup4==4.13.4
chainlit==2.7.1.1
python-dotenv==1.0.0
